Reactify - Add useToken hook
